+++ common/src/main/java/com/samsthenerd/monthofswords/registry/SwordsModItems.java
@@ -37,7 +37,7 @@ public class SwordsModItems {
     // make our creative tab.
     public static final RegistrySupplier<ItemGroup> SWORDS_MOD_GROUP = TABS.register("monthofswords_tab", () ->
             CreativeTabRegistry.create(Text.translatable("itemgroup.monthofswords.general"),
-                    () -> new ItemStack(WING_SWORD.get())));
+                    () -> new ItemStack(CURSED_SWORD.get())));
 
     /*
      * Helper function for registering an item.