+++ changelog.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## 0.3.0 (Initial Release)
+## 0.4.0 (Initial Release)
 
 ### Added 
 
-- Cursed Sword
-- Winged Sword
+- Divine Sword
+- Cursed Sword (not yet available in survival)
 - Floral Sword
+- Winged Sword